Fixed several bugs in Thorium
🐛 Bug Fixes

- *(operator)* Fixed issue where the config could not be made into a CRD
- *(scaler)* Scalers now only requests details on clusters they care about
- *(agent)* Fixed issue where the agent was not injecting kwargs correctly
- *(api)* Fixed issue where the API was incorrectly rejecting result paths

Change Details

fix(agent): Fixed issue where the agent was not injecting kwargs correctly

This was causing the agent to add a list of values after each kwargs
instead of repeating the kwarg for each value. This means Thorium will
now use --kwarg <value> --kwarg <value> instead of --kwarg <value>
<value>.

fix(api): Fixed issue where the API was incorrectly rejecting result paths

This was due to an incorrect check for '..' in file paths.

fix(operator): Fixed issue where the config could not be made into a CRD

This was caused by an enum having different types for each branch. The
downside of this fix is that our config does allow someone to configure
certificate validation settings while also disabling certificate
validation. That could lead to some confusing scenarios where you think
validation is enabled but its not.

fix(scaler): Scalers now only requests details on clusters they care about

This helps resolves issues where the scaler tries to get info on clusters
that it cannot and will not schedule on.

Closes #31